requirements.txt
graia-ariadne==0.5.3.post3